test(nhibernate): extract common integration test base class
- Extract common NHibernate test setup logic into new NHibernateTestBase class
- Move GetNHibernateConfiguration, CreateAndClearPersonsTable methods to base class for reuse across tests
- Update AuroraConnectionTrackerConnectivityTests, AwsWrapperDriverTests and CustomEndpointConnectivityTests to inherit from NHibernateTestBase
